+++ core/class/snips.assistant.manager.class.php
@@ -6,59 +6,38 @@ require_once dirname(__FILE__) . '/snips.utils.class.php';
 
 class SnipsAssistantManager
 {
-    /* create snips-intent object */
-    static function create_snips_object($assistant)
-    {
-        // since 3.3.3, use jeeObject instead of object as class name
-        if (version_compare(jeedom::version(), '3.3.3', '>=')) {
-            $obj_field = 'jeeObject';
-            SnipsUtils::logger('Jeedom >= 3.3.3');
-        }else{
-            $obj_field = 'object';
-            SnipsUtils::logger('Jeedom <= 3.3.3');
-        }
-
-        $obj = object::byName('Snips-Intents');
-        if (!isset($obj) || !is_object($obj)) {
-            $obj = new $obj_field();
-            $obj->setName('Snips-Intents');
-            SnipsUtils::logger('Created object: Snips-Intents');
-        }
-        $obj->setIsVisible(1);
-        $obj->setConfiguration('id', $assistant["id"]);
-        $obj->setConfiguration('name',$assistant["name"]);
-        $obj->setConfiguration('hotword', $assistant['hotword']);
-        $obj->setConfiguration('language', $assistant['language']);
-        $obj->setConfiguration('createdAt', $assistant['createdAt']);
-        $obj->save();
-    }
-
     /* create intent eq objects */
     static function create_intents_eq_cmd($raw_intents)
     {
-        $object_id = object::byName('Snips-Intents')->getId();
-
+        $count = 0;
+        $object_id = SnipsUtils::get_snips_intent_object()->getId();
         foreach ($raw_intents as $intent) {
-            if (strpos(strtolower($intent['name']), 'jeedom')) {
-                $elogic = snips::byLogicalId($intent['id'], 'snips');
-                if (!is_object($elogic)) {
-                    $elogic = new snips();
-                    $elogic->setLogicalId($intent['id']);
-                    $elogic->setName($intent['name']);
-                    SnipsUtils::logger('Created intent entity: '.$intent['name']);
-                }
-                $elogic->setEqType_name('snips');
-                $elogic->setIsEnable(1);
-                $elogic->setConfiguration('snipsType', 'Intent');
-                $elogic->setConfiguration('slots', $intent['slots']);
-                $elogic->setConfiguration('isSnipsConfig', 1);
-                $elogic->setConfiguration('isInteraction', 0);
-                $elogic->setConfiguration('language', $intent['language']);
-                $elogic->setObject_id($object_id);
-                $elogic->save();
+            if (!strpos(strtolower($intent['name']), 'jeedom')) {
+                continue;
             }
+            // create intent
+            $elogic = snips::byLogicalId($intent['id'], 'snips');
+            if (!is_object($elogic)) {
+                $elogic = new snips();
+                $elogic->setLogicalId($intent['id']);
+                $elogic->setName($intent['name']);
+                SnipsUtils::logger('Created intent entity: '.$intent['name']);
+            }
+            $elogic->setEqType_name('snips');
+            $elogic->setIsEnable(1);
+            $elogic->setConfiguration('snipsType', 'Intent');
+            $elogic->setConfiguration('slots', $intent['slots']);
+            $elogic->setConfiguration('isSnipsConfig', 1);
+            $elogic->setConfiguration('isInteraction', 0);
+            $elogic->setConfiguration('language', $intent['language']);
+            $elogic->setObject_id($object_id);
+            $elogic->save();
+            // generate all the necessary command for this intent
             self::create_slots_cmd($elogic, $intent['slots']);
+            $count += 1;
         }
+
+        SnipsUtils::logger('loaded '. $count .' jeedom intent.', 'info');
     }
 
     /* create slots cmd for all the intent eq objects */
@@ -91,7 +70,7 @@ class SnipsAssistantManager
         $json_string = file_get_contents($assistant_file);
         $assistant = json_decode($json_string, true);
 
-        self::create_snips_object($assistant);
+        SnipsUtils::create_snips_intent_object($assistant);
 
         self::create_intents_eq_cmd($assistant['intents']);
 
@@ -137,7 +116,7 @@ class SnipsAssistantManager
 
     /* create snips tts eq object */
     static function create_snips_device($site_id, $lang){
-        $obj = object::byName('Snips-Intents');
+        $obj = SnipsUtils::get_snips_intent_object();
         if (!$obj || !$site_id) {
             return;
         }